# Replace greedy completer with guarded evaluation, improve dict completions