… .github/workflows/mypy.yml
… IPython/core/magics/config.py
… IPython/core/tests/test_guarded_eval.py