Addresses https://github.com/ipython/ipython/issues/13735#issuecomment-1306121425 and https://github.com/ipython/ipython/issues/13836.

### Guarded evaluation

A new module with `guarded_eval()` function is added offering guarded evaluation of Python code. It uses the built-in `ast` parser (no dependencies), similarly to existing solutions:

<details>

- `ast.literal_eval` - too restricted, cannot handle most expressions that we need
- [`asteval`](https://github.com/newville/asteval/) - is close but aimed at a different use case

</details>

The purpose of `guarded_eval()` is to guard against unwanted side-effects only, no further guarantees are made.

<details>

`guarded_eval` does not evaluate actions which always have side effects:
  - class definitions (``class sth: ...``)
  - function definitions (``def sth: ...``)
  - variable assignments (``x = 1``)
  - augmented assignments (``x += 1``)
  - deletions (``del x``)

It also does not evaluate operations which do not return values:
  - assertions (``assert x``)
  - pass (``pass``)
  - imports (``import x``)
  - control flow
     - conditionals (``if x:``) except for ternary IfExp (``a if x else b``)
     - loops (``for`` and `while``)
     - exception handling

</details>

Since other parts of the codebase call bare `eval` I would like to follow-up replacing these calls with safer `guarded_eval` across the board.

### Better dict key completions

Enhancements to dictionary key completions:
- completions of nested dictionaries will work by default
   - the restriction which was implemented using regular expression followed by unsafe `eval()` was lifted
   - the use of `eval()` was replaced with safe `guarded_eval()`
   - by default this only works for built-in `__getattr__`; if user-defined `__getattr__` or `__getitem__` gets encountered on the evaluation path, the evaluation will be aborted to avoid triggering side effects
   - evaluation policy can be adjusted via configuration (e.g. forbidding evaluation or relaxing the rules)
   ```python
   d = {'a': {'b': None}}
   d['a'][<tab>   # will suggest 'b'
   l = [d]
   l[0][<tab>     # will suggest 'a'
   ```
- add support for pandas columns when using `.loc` indexer
   ```python
   df.loc[:, <tab>    # will suggest column names
   ```
- add support for numbers
  ```python
   d = {999: 'emergency'}
   d[9<tab>    # will suggest `999`
   ```
- support auto-closing of strings and brackets outside of greedy completer
- add support for proper auto-closing of tuple key items
  ```python
  d = {('aa', 'bb'): None}
  d[<tab>   # will complete to `d['aa', `
  ```

### Changes in user config options

#### `Completer.evaluation` replaces `Completer.greedy`

Successive options allow to enable more eager evaluation for more accurate completion suggestions,
including for nested dictionaries, nested lists, or even results of function calls. Setting `unsafe`
or higher can lead to evaluation of arbitrary user code on TAB with potentially dangerous side effects.
Allowed values are:
  - `forbidden`: no evaluation at all
  - `minimal`:  evaluation of literals and access to built-in namespaces; no item/attribute evaluation nor access to locals/globals
  - `limited` (default): access to all namespaces, evaluation of hard-coded methods (`keys()`, `__getattr__`, `__getitems__`, etc) on allow-listed objects (e.g. `dict`, `list`, `tuple`, `pandas.Series`)
  - `unsafe`: evaluation of all methods and function calls but not of syntax with side-effects like `del x`,
  - `dangerous`: completely arbitrary evaluation

#### `Completer.auto_close_dict_keys` is added

Previously closing of strings and brackets was only ever active in greedy completer. Now the logic was expanded to handle keys of tuples correctly; this setting allows to enable auto-closing regardless of evaluation policy.

#### `Completer.greedy` is deprecated

When enabled in  `Completer.greedy` activates following settings for compatibility:
  - `evaluation = 'unsafe'`
  - `auto_close_dict_keys = True`

While all old tests for greedy completer pass with these settings, the old greedy completer was essentially running on `dangerous` mode. I am not sure if we should be adding it, but I wanted to keep it so users can change to `evaluation = 'dangerous'` at least in the next version in case if we need to fix something with `guarded_eval`.


### Remaining tasks

- [x] polish documentation
- [x] add evaluation guards against other dunder methods
- [x] add more tests
- [x] benchmark performance to ensure that this does not introduce performance regressions

NTH:
- [ ] maybe port attribute matcher to remove greedy completer remnants (up for discussion as to avoid breaking downstream)
- [ ] port dict key matcher to v2 and return "column" type for pandas rather than "dict key"; what is the right type name for numpy?