@krassowski 2022-12-04
Averaging across all scenarios this PR has a small overall performance impact of 6.41% (from grand average of 11.29 ms to 12.01 ms; before refactor the average would have been 11.72 ms). I would argue that this is ok given the added features.

Breakdown by parameter (each scenario has 5 runs x # other parameter combinations; mean [1st - 3rd quartile]):

| param     | value       | main                     | this PR                  |
|:----------|:------------|:-------------------------|:-------------------------|
| greedy    | 0           | 56.36 [5.74 - 101.93]    | 59.40 [5.66 - 104.03]    |
| greedy    | 1           | 56.55 [5.79 - 103.30]    | 60.75 [5.69 - 103.33]    |
| n_tokens  | 5           | 51.42 [4.11 - 97.68]     | 56.89 [5.39 - 102.51]    |
| n_tokens  | 10          | 51.75 [5.11 - 102.20]    | 53.63 [5.45 - 103.61]    |
| n_tokens  | 100         | 56.89 [5.58 - 111.42]    | 58.68 [5.65 - 120.21]    |
| **n_tokens**  | **1000**        | **65.78 [9.15 - 127.22]**    | **71.12 [17.60 - 120.69]**   |
| nested_n  | 0           | 59.06 [5.76 - 100.31]    | 61.38 [5.69 - 104.28]    |
| nested_n  | 1           | 55.29 [5.77 - 102.40]    | 60.07 [5.66 - 103.40]    |
| nested_n  | 2           | 54.68 [5.76 - 98.67]     | 59.23 [5.65 - 103.10]    |
| nested_n  | 3           | 55.72 [5.74 - 101.00]    | 59.71 [5.73 - 104.19]    |
| nested_n  | 4           | 56.17 [5.79 - 103.14]    | 59.26 [5.65 - 101.84]    |
| nested_n  | 5           | 57.83 [5.73 - 103.63]    | 60.81 [5.72 - 105.13]    |
| query     |             | 91.10 [86.45 - 94.05]    | 94.56 [87.68 - 96.54]    |
| query     | [           | 123.56 [103.76 - 138.67] | 130.70 [106.44 - 154.95] |
| query     | ["1         | 5.09 [2.88 - 6.35]       | 8.99 [3.23 - 10.98]      |
| query     | ["1111111"] | 6.08 [5.76 - 6.31]       | 6.07 [5.65 - 6.08]       |
| token_len | 5           | 57.48 [5.77 - 102.59]    | 60.89 [5.66 - 103.15]    |
| token_len | 10          | 56.97 [5.79 - 102.67]    | 59.53 [5.69 - 103.82]    |
| token_len | 25          | 54.92 [5.73 - 101.39]    | 59.82 [5.68 - 104.36]    |

The case with very long tokens (keys) in the dictionary appears worth further investigation (though I think that it should not hold this PR as it is rather uncommon to have so long dictionary keys).

<details>

```python
import timeit
import subprocess

import pandas as pd
from tqdm.auto import tqdm

import IPython
from IPython import get_ipython


def time_dict_completions(
    token_len = 10,
    access_key_len=10,
    n_tokens = 100,
    dict_name_len = 10,
    query = '[',
    nested_n=0,
    greedy=False
):
    ip = get_ipython()
    dict_name = "x" * dict_name_len

    leaf_dict = {
        f'{i}'.ljust(token_len, 'x'): None
        for i in range(n_tokens)
    }
    top_dict = {}
    current_dict = top_dict

    access_key = 'b' * access_key_len  # b as in branch
    for i in range(nested_n):
        current_dict[access_key] = {}

    current_dict.update(leaf_dict)

    ip.user_ns[dict_name] = top_dict
    ip.Completer.greedy = greedy
    line = f'{dict_name}{query}'

    return timeit.timeit('ip.complete(line)', globals=locals(), number=5)


def test_grid():
    return [
        dict(
            n_tokens=n_tokens,
            token_len=token_len,
            nested_n=nested_n,
            query=query,
            greedy=greedy
        )
        for greedy in [True, False]
        for n_tokens in [5, 10, 100, 1000]
        for token_len in [5, 10, 25]
        for nested_n in [0, 1, 2, 3, 4, 5]
        for query in ['[', '', '["1', '["1111111"]']
    ]

if __name__ == '__main__':

    result = pd.DataFrame([
        {
            'time': time_dict_completions(**opts),
            **opts
        }
        for opts in tqdm(test_grid())
    ])
    commit = (
        subprocess
        .check_output(['git', 'rev-parse', '--short', 'HEAD'])
        .decode('ascii')
        .strip()
    )

    result.to_csv(f'dict_completions_times_{IPython.__version__}_{commit}.csv')
```

</details>