@krassowski 2022-12-04
The coverage failure is unrelated; this PR should increase coverage but for some time now the coverage reports are incorrect (https://github.com/ipython/ipython/pull/13853) for which I sent a fix in https://github.com/ipython/ipython/issues/13854.