@aldanor 2023-03-14
@krassowski Wonder how this is supposed to be used (e.g. by library developers)?

While it's nice that numpy / pandas are hardcoded in, is it possible to modify the selective policy at runtime from within the kernel? (e.g. to allow eager getattr/getitem on some `foo.bar.Baz` type?)