@krassowski 2023-03-14
@aldanor this is a good question. One could swap policies in `EVALUATION_POLICIES` but this would not be convenient for library authors. I think it would be fine to expose a function allowing to add a custom allowed properties. Would you mind opening a new issue with reference to this PR and your thoughts on how you would see it?