@aldanor 2023-03-14
I kind of got the point of something like this
```python
from IPython.core import guarded_eval
guarded_eval.EVALUATION_POLICIES['limited'].allowed_getattr_external.update(types)
guarded_eval.EVALUATION_POLICIES['limited'].allowed_getitem_external.update(types)
guarded_eval.SUPPORTED_EXTERNAL_GETITEM.update(types)
```
but this feels extremely fiddle and fragile and I'm not sure it fully works 🙂

---

Also, I can't seem to be able to make this work:
```python
class X:
    @property
    def y(self) -> Y: ...

class Y:
    def __getitem__(self, index): ...
    def _ipython_key_completions_(self): ...
```

Something like this triggers getitem completion just fine
```python
>>> y = x.y
>>> y['
```
whereas this doesn't
```python
>>> x.y['
```
and so I started digging to see if it's possible by modifying guarded eval policies and white-listing all of the above and it doesn't seem to solve it either 🤔 (not sure it's directly related but perhaps it is, so I figured I'd mention the use case)

---

> Would you mind opening a new issue with reference to this PR and your thoughts on how you would see it?

I guess I could, once I figure how to solve my use case so I have an idea of how it works (could be something like a `guarded_eval.whitelist(user_type, **kwargs)`).