@aldanor 2023-03-14
Update: the above seems to have been resolved by adding the type `X` directly to limited policy's `allowed_getattr` set. Not sure why it doesn't work with `allowed_getattr_external` though.